tests/wpt/tests/css/css-text/parsing/text-align-last-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-align-last with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-align-last">
<meta name="assert" content="text-align-last supports only the grammar 'auto | start | end | left | right | center | justify | match-parent'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("text-align-last", "none");
test_invalid_value("text-align-last", "auto start");
test_invalid_value("text-align-last", "end left");
test_invalid_value("text-align-last", "right center");
test_invalid_value("text-align-last", "justify match-parent");
</script>
</body>
</html>
